+++ NEWS.md
@@ -1,4 +1,4 @@
-# pkgdown (development version)
+# pkgdown 2.0.1
 
 * Fix CRAN failures.